+++ reports/coursedashboard/lang/de/lareport_coursedashboard.php
@@ -24,6 +24,8 @@
 
 $string['pluginname'] = 'Learning Analytics Dashboard';
 
+$string['privacy:metadata'] = 'Dieses Plugin speichert keine personenbezogenen Daten.';
+
 $string['registered_users'] = 'Teilnehmer/innen';
 $string['active_learners'] = 'Aktive Nutzer';
 $string['click_count'] = 'Aufrufe';